refactor: move felt from common to types
