+++ pkg/types/felt_test.go
@@ -1,4 +1,4 @@
-package common
+package types
 
 import (
 	"bytes"
